curso-1/challenge-01/pyproject.toml
[project]
name = "challenge-01-hello"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "python-dotenv",
]
